+++ xash3d_engine/engine/client/mod_dbghulls.c
@@ -135,7 +135,7 @@ static winding_t* winding_for_plane(const mplane_t* p)
 
 	for ( i = 0; i < 3; i++ )
 	{
-		v = fabs(p->normal[i]);
+		v = fabsf(p->normal[i]);
 		if ( v > max )
 		{
 			axis = i;
@@ -557,7 +557,7 @@ static void hull_windings_r(hull_t* hull, mclipnode_t* node, hullnode_t* polys, 
 	hullnode_t frontlist = LIST_HEAD_INIT(frontlist);
 	hullnode_t backlist = LIST_HEAD_INIT(backlist);
 	winding_t *w, *next, *front, *back;
-	int i;
+	uint i;
 
 	list_for_each_entry_safe(w, next, polys, chain)
 	{
@@ -607,7 +607,7 @@ static void hull_windings_r(hull_t* hull, mclipnode_t* node, hullnode_t* polys, 
 	for ( i = 0; w && i < node_stack_depth; i++ )
 	{
 		mplane_t* p = hull->planes + node_stack[i]->planenum;
-		w = winding_clip(w, p, false, side_stack[i], 0.00001);
+		w = winding_clip(w, p, false, side_stack[i], 0.00001f);
 	}
 
 	if ( w )
